proxy/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.1.0
backoff==2.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==43.0.3
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-sso==0.16.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.60.0
MarkupSafe==3.0.2
multidict==6.1.0
oauthlib==3.2.2
openai==1.61.1
orjson==3.10.15
packaging==24.2
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.0.1
python-multipart==0.0.18
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
rq==2.1.0
sniffio==1.3.1
starlette==0.45.3
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.29.0
uvloop==0.21.0
yarl==1.18.3
zipp==3.21.0
